structure: move settings route into directory
